Renamed files to live in proper subdirectory relative to galaxy root
There are a number of files that will be modified in a "default" galaxy
installation to support MCell. The current plan for this repository is
that it contains only those files that have been modified from the
defatult galaxy installation. So copying this directory structure on top
of a fresh galaxy installation should mostly work.