Added a .gitignore file to ignore the mcell executable
The current MCell/Galaxy development environment contains a link:

  galaxy/tools/mcell -> mcell_galaxy/tools/mcell

This uses the git versions of all .xml and .py files automatically
without having to copy them. However, that directory also needs a copy
of mcell to run (which is not in the repository). To keep things simple,
the mcell executable resides in the git working directory (with the
other files) and this .gitignore file keeps git from reporting it.