Added a Show Data Model tool to display the parameters and raw fields.
It's not clear how to have parameters in a data model displayed as edit
fields in Galaxy. However, it is fairly easy (already partially in the
MCell tool) to add name/value pairs to be modified. So one approach
would be to allow data model modification with a combination of the two
tools already mentioned. The show data model would show the available
parameters, and then a separate application could modify the data model
(creating a new instance in Galaxy) as needed.