Added ability to modify selected parameters by name.
This interface requires knowing the names of parameters in advance. This
can be obtained by running the "Show Data Model" tool which will display
an HTML output of all parameter names, values, units, and descriptions
for the chosen data model file. It would be nice to be able to present
these in the modification tool, but that hasn't been figured out yet.