Added a test bed tool for testing out Galaxy tool options.
The test bed code explores the XML tags in mostly the order that they're
listed in the documentation.

There are two options that might be especially helpful:

   tool > inputs > param > options (from_dataset option)
   tool > outputs > data > discover_datasets

The first appears to be able to create user interface items from the
contents of a data set. The second provides a way to use regular
expressions to "harvest" output files produced by a job and create
Galaxy data sets from them.