More work on dynamic options to allow easier setting of parameters.
These updates reflect some web research on how to add dynamic options to
a Galaxy tool's input section. So far, the best result is a "repeat"
block that allows the user to add new parameter change blocks and
specify both the name and value of each parameter to be modified. The
dyn_opts.xml/.py files isolated this approach for presentation to the
biostar.usegalaxy.org forum page:

   https://biostar.usegalaxy.org/p/20309/

At this time there has been no response.